Expose Journal-related files to allow external Journal implementation integration.
