+++ accord-core/src/main/java/accord/impl/InMemoryCommandStore.java
@@ -43,18 +43,16 @@ import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
 import com.google.common.annotations.VisibleForTesting;
-
-import accord.api.LocalListeners;
-import accord.api.RoutingKey;
-import accord.impl.progresslog.DefaultProgressLog;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import accord.api.Agent;
 import accord.api.DataStore;
+import accord.api.Journal;
+import accord.api.LocalListeners;
 import accord.api.ProgressLog;
-import accord.local.Cleanup;
+import accord.api.RoutingKey;
+import accord.impl.progresslog.DefaultProgressLog;
 import accord.local.Command;
 import accord.local.CommandStore;
 import accord.local.CommandStores.RangesForEpoch;
@@ -97,19 +95,17 @@ import static accord.local.SafeCommandStore.TestDep.ANY_DEPS;
 import static accord.local.SafeCommandStore.TestDep.WITH_OR_INVALIDATED;
 import static accord.local.SafeCommandStore.TestStartedAt.STARTED_BEFORE;
 import static accord.local.SafeCommandStore.TestStatus.ANY_STATUS;
+import static accord.primitives.Routables.Slice.Minimal;
 import static accord.primitives.SaveStatus.Applying;
 import static accord.primitives.SaveStatus.Erased;
 import static accord.primitives.SaveStatus.ErasedOrVestigial;
 import static accord.primitives.SaveStatus.ReadyToExecute;
 import static accord.primitives.Status.Applied;
 import static accord.primitives.Status.Durability.Local;
-import static accord.primitives.Status.Invalidated;
-import static accord.primitives.Status.PreApplied;
+import static accord.primitives.Status.NotDefined;
 import static accord.primitives.Status.PreCommitted;
 import static accord.primitives.Status.Stable;
 import static accord.primitives.Status.Truncated;
-import static accord.primitives.Status.NotDefined;
-import static accord.primitives.Routables.Slice.Minimal;
 import static accord.utils.Invariants.illegalState;
 import static java.lang.String.format;
 
@@ -130,10 +126,12 @@ public abstract class InMemoryCommandStore extends CommandStore
     protected Timestamp maxRedundant = Timestamp.NONE;
 
     private InMemorySafeStore current;
+    private final Journal.Loader loader;
 
     public InMemoryCommandStore(int id, NodeCommandStoreService time, Agent agent, DataStore store, ProgressLog.Factory progressLogFactory, LocalListeners.Factory listenersFactory, EpochUpdateHolder epochUpdateHolder)
     {
         super(id, time, agent, store, progressLogFactory, listenersFactory, epochUpdateHolder);
+        this.loader = new CommandLoader(this);
     }
 
     protected boolean canExposeUnloaded()
@@ -1038,22 +1036,22 @@ public abstract class InMemoryCommandStore extends CommandStore
         public void postExecute()
         {
             commands.values().forEach(c -> {
-                if (c != null && c.current() != null)
-                {
-                    Timestamp executeAt = c.current().executeAtIfKnown();
-                    if (executeAt == null)
-                        return;
+                if (c == null || c.current() == null)
+                    return;
 
+                Timestamp executeAt = c.current().executeAtIfKnown();
+                if (executeAt != null)
+                {
                     if (c.current().hasBeen(Truncated)) commandStore.commandsByExecuteAt.remove(executeAt);
                     else commandStore.commandsByExecuteAt.put(executeAt, commandStore.command(c.txnId()));
                 }
-            });
 
-            commands.values().forEach(c -> {
-                if (c.isUnset())
+                if (c.isUnset() || c.current().saveStatus().isUninitialised())
                     commandStore.commands.remove(c.txnId());
+
                 c.invalidate();
             });
+
             timestampsForKey.values().forEach(tfk -> {
                 if (tfk.isUnset())
                     commandStore.timestampsForKey.remove(tfk.key());
@@ -1377,99 +1375,88 @@ public abstract class InMemoryCommandStore extends CommandStore
         unsafeSetRejectBefore(new RejectBefore());
     }
 
-    public interface Loader
+    public Journal.Loader loader()
     {
-        void load(Command next);
-        void apply(Command next);
+        return loader;
     }
 
-    public Loader loader()
+    private static class CommandLoader extends AbstractLoader
     {
-        return new Loader()
+        private final InMemoryCommandStore commandStore;
+
+        private CommandLoader(InMemoryCommandStore commandStore)
         {
-            private PreLoadContext context(Command command, KeyHistory keyHistory)
+            this.commandStore = commandStore;
+        }
+
+        private PreLoadContext context(Command command, KeyHistory keyHistory)
+        {
+            TxnId txnId = command.txnId();
+            AbstractUnseekableKeys keys = null;
+
+            if (CommandsForKey.manages(txnId))
+                keys = (AbstractUnseekableKeys) command.participants().hasTouched();
+            else if (!CommandsForKey.managesExecution(txnId) && command.hasBeen(Status.Stable) && !command.hasBeen(Status.Truncated))
+                keys = command.asCommitted().waitingOn.keys;
+
+            if (keys != null)
             {
-                TxnId txnId = command.txnId();
-                AbstractUnseekableKeys keys = null;
-
-                if (CommandsForKey.manages(txnId))
-                    keys = (AbstractUnseekableKeys) command.participants().hasTouched();
-                else if (!CommandsForKey.managesExecution(txnId) && command.hasBeen(Status.Stable) && !command.hasBeen(Status.Truncated))
-                    keys = command.asCommitted().waitingOn.keys;
-
-                if (keys != null)
-                {
-                    return PreLoadContext.contextFor(txnId, keys, keyHistory);
-                }
-
-                return PreLoadContext.contextFor(txnId);
+                return PreLoadContext.contextFor(txnId, keys, keyHistory);
             }
 
-            public void load(Command command)
+            return PreLoadContext.contextFor(txnId);
+        }
+
+        @Override
+        public void load(Command command, Journal.OnDone onDone)
+        {
+            try
             {
-                TxnId txnId = command.txnId();
-
-                executeInContext(InMemoryCommandStore.this,
-                                 context(command, ASYNC),
-                                 safeStore -> {
-                                     Command local = command;
-                                     if (local.status() != Truncated && local.status() != Invalidated)
-                                     {
-                                         Cleanup cleanup = Cleanup.shouldCleanup(safeStore, local, local.participants());
-                                         switch (cleanup)
-                                         {
-                                             case NO:
-                                                 break;
-                                             case INVALIDATE:
-                                             case TRUNCATE_WITH_OUTCOME:
-                                             case TRUNCATE:
-                                             case ERASE:
-                                                 local = Commands.purge(local, local.participants(), cleanup);
-                                         }
-                                     }
-
-                                     local = safeStore.unsafeGet(txnId).update(safeStore, local);
-                                     if (local.status() == Truncated)
-                                         safeStore.progressLog().clear(local.txnId());
-                                     return local;
-                                 });
-
-
+                commandStore.executeInContext(commandStore,
+                                              context(command, ASYNC),
+                                              safeStore -> loadInternal(command, safeStore));
+                onDone.success();
             }
-
-            public void apply(Command command)
+            catch (Throwable t)
             {
-                TxnId txnId = command.txnId();
+                onDone.failure(t);
+            }
+        }
 
+        @Override
+        public void apply(Command command, Journal.OnDone onDone)
+        {
+            try
+            {
                 PreLoadContext context = context(command, KeyHistory.TIMESTAMPS);
-                executeInContext(InMemoryCommandStore.this,
-                                 context,
-                                 safeStore -> {
-                                     SafeCommand safeCommand = safeStore.unsafeGet(txnId);
-                                     Command local = safeCommand.current();
-                                     if (local.is(Stable) || local.is(PreApplied))
-                                     {
-                                         Commands.maybeExecute(safeStore, safeCommand, local, true, true);
-                                     }
-                                     else if (local.saveStatus().compareTo(Applying) >= 0 && !local.hasBeen(Truncated))
-                                     {
-                                         unsafeApplyWrites(safeStore, safeCommand, local);
-                                     }
-                                     return null;
-                                 });
+                commandStore.executeInContext(commandStore,
+                                              context,
+                                              safeStore -> {
+                                                  applyWrites(command, safeStore, (safeCommand, cmd) -> {
+                                                      unsafeApplyWrites(safeStore, safeCommand, cmd);
+                                                  });
+                                                  return null;
+                                              });
             }
-        };
-    }
+            catch (Throwable t)
+            {
+                onDone.failure(t);
+                return;
+            }
 
-    public static void unsafeApplyWrites(SafeCommandStore safeStore, SafeCommand safeCommand, Command command)
-    {
-        Command.Executed executed = command.asExecuted();
-        Participants<?> executes = executed.participants().executes(safeStore, command.txnId(), command.executeAt());
-        if (!executes.isEmpty())
+            onDone.success();
+        }
+
+        protected void unsafeApplyWrites(SafeCommandStore safeStore, SafeCommand safeCommand, Command command)
         {
-            command.writes().applyUnsafe(safeStore, Commands.applyRanges(safeStore, command.executeAt()), command.partialTxn());
-            safeCommand.applied(safeStore);
-            safeStore.notifyListeners(safeCommand, command);
+            Command.Executed executed = command.asExecuted();
+            Participants<?> executes = executed.participants().executes(safeStore, command.txnId(), command.executeAt());
+            if (!executes.isEmpty())
+            {
+                command.writes().applyUnsafe(safeStore, Commands.applyRanges(safeStore, command.executeAt()), command.partialTxn());
+                safeCommand.applied(safeStore);
+                safeStore.notifyListeners(safeCommand, command);
+            }
         }
     }
