+++ accord-core/src/test/java/accord/Utils.java
@@ -18,30 +18,28 @@
 
 package accord;
 
-import java.time.Duration;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Set;
+import java.util.TreeMap;
 import java.util.TreeSet;
-import java.util.concurrent.TimeUnit;
 import java.util.function.Function;
 
-import accord.api.Agent;
 import com.google.common.collect.Sets;
 
+import accord.api.Agent;
 import accord.api.Key;
+import accord.api.LocalConfig;
 import accord.api.MessageSink;
 import accord.api.Scheduler;
 import accord.api.TopologySorter;
-import accord.api.LocalConfig;
 import accord.coordinate.CoordinationAdapter;
+import accord.impl.DefaultLocalListeners;
+import accord.impl.DefaultRemoteListeners;
 import accord.impl.DefaultTimeouts;
 import accord.impl.InMemoryCommandStores;
 import accord.impl.IntKey;
-import accord.impl.DefaultLocalListeners;
-import accord.impl.progresslog.DefaultProgressLogs;
-import accord.impl.DefaultRemoteListeners;
 import accord.impl.SizeOfIntersectionSorter;
 import accord.impl.TestAgent;
 import accord.impl.list.ListQuery;
@@ -50,6 +48,7 @@ import accord.impl.list.ListUpdate;
 import accord.impl.mock.MockCluster;
 import accord.impl.mock.MockConfigurationService;
 import accord.impl.mock.MockStore;
+import accord.impl.progresslog.DefaultProgressLogs;
 import accord.local.DurableBefore;
 import accord.local.Node;
 import accord.local.Node.Id;
@@ -67,8 +66,6 @@ import accord.utils.EpochFunction;
 import accord.utils.Invariants;
 import accord.utils.SortedArrays.SortedArrayList;
 import accord.utils.ThreadPoolScheduler;
-import org.awaitility.Awaitility;
-import org.awaitility.core.ThrowingRunnable;
 
 import static accord.utils.async.AsyncChains.awaitUninterruptibly;
 
@@ -132,9 +129,10 @@ public class Utils
 
     public static Txn listWriteTxn(Id client, Keys keys)
     {
-        ListUpdate update = new ListUpdate(Function.identity());
+        TreeMap<Key, Integer> map = new TreeMap<>();
         for (Key k : keys)
-            update.put(k, 1);
+            map.put(k, 1);
+        ListUpdate update = new ListUpdate(Function.identity());
         ListRead read = new ListRead(Function.identity(), false, keys, keys);
         ListQuery query = new ListQuery(client, keys.size(), false);
         return new Txn.InMemory(keys, read, query, update);
@@ -200,21 +198,6 @@ public class Utils
                              localConfig);
         awaitUninterruptibly(node.unsafeStart());
         return node;
-    }
-
-    public static void spinUntilSuccess(ThrowingRunnable runnable)
-    {
-        spinUntilSuccess(runnable, 10);
-    }
-
-    public static void spinUntilSuccess(ThrowingRunnable runnable, int timeoutInSeconds)
-    {
-        Awaitility.await()
-                  .pollInterval(Duration.ofMillis(100))
-                  .pollDelay(0, TimeUnit.MILLISECONDS)
-                  .atMost(timeoutInSeconds, TimeUnit.SECONDS)
-                  .ignoreExceptions()
-                  .untilAsserted(runnable);
     }
 
     public static TopologyManager testTopologyManager(TopologySorter.Supplier sorter, Id node)
